+++ Sources/Alchemy/ORM/Database/ConnectionPool.swift
@@ -1,0 +1,5 @@
+import PostgresKit
+
+// Vapor's `AsyncKit` has `EventLoopGroupConnectionPool` with a generic param. When we do our own we won't
+// have that. For now leaving as `EventLoopGroupConnectionPool<PostgresConnectionSource>`.
+public typealias ConnectionPool = EventLoopGroupConnectionPool<PostgresConnectionSource>
